+++ example/lib/tabs/yt_web.dart
@@ -47,7 +47,7 @@ class YTWeb extends StatelessWidget {
             ),
             playerTheme: OmniVideoPlayerThemeData().copyWith(
               icons: VideoPlayerIconTheme().copyWith(error: Icons.warning),
-              overlays: VideoPlayerOverlayTheme().copyWith(
+              backdrop: VideoPlayerBackdropTheme().copyWith(
                 backgroundColor: Colors.white,
                 alpha: 100,
               ),