+++ lib/src/widgets/controls/playback_speed_menu_button.dart
@@ -28,8 +28,16 @@ class PlaybackSpeedMenuButton extends StatelessWidget {
     return Card(
       elevation: 8,
       color: theme.colors.menuBackground,
-      child: SizedBox(
+      child: Container(
         width: 110,
+        decoration:
+            theme.menus.menuDecoration ??
+            BoxDecoration(
+              color: theme.colors.menuBackground,
+              borderRadius: BorderRadius.circular(
+                theme.shapes.menuBorderRadius,
+              ),
+            ),
         child: ListView(
           padding: const EdgeInsets.symmetric(vertical: 8),
           shrinkWrap: true,